Issue #96
