+++ tests/data/test_demonstration_time_series.py
@@ -1,6 +1,22 @@
 
 import numpy as np
-from dtaianomaly.data.synthetic import make_sine_wave, inject_anomalies
+from dtaianomaly.data.demonstration_time_series import make_sine_wave, inject_anomalies, demonstration_time_series, DemonstrationTimeSeriesLoader
+
+
+class TestDemonstrationTimeSeriesLoader:
+
+    def test(self):
+        dataset = DemonstrationTimeSeriesLoader().load()
+        X, y = demonstration_time_series()
+        assert np.array_equal(X, dataset.X_test)
+        assert np.array_equal(y, dataset.y_test)
+        assert dataset.X_train is None
+        assert dataset.y_train is None
+
+    def test_str(self):
+        assert str(DemonstrationTimeSeriesLoader()) == 'DemonstrationTimeSeriesLoader()'
+        assert str(DemonstrationTimeSeriesLoader(False)) == 'DemonstrationTimeSeriesLoader()'
+        assert str(DemonstrationTimeSeriesLoader(True)) == 'DemonstrationTimeSeriesLoader(do_caching=True)'
 
 
 class TestInjectAnomalies:
